pyproject.toml
[project]
name = "confme"
version = "2.0.4"
description = "Easy configuration management in python"
authors = [{ name = "Iwan Silvan Bolzern", email = "iwan.bolzern@roche.com" }]
requires-python = ">=3.10,<4"
readme = "README.md"
license = "MIT"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "pyyaml>=5.3",
    "pydantic>=2.12,<3",
    "tabulate>=0.8.9",
    "typing-extensions>=4.11.0,<5",
]

[project.urls]
Homepage = "https://github.com/iwanbolzern/confme"
Repository = "https://github.com/iwanbolzern/confme"

[dependency-groups]
dev = [
    "pytest>=6.1.2",
    "pyright ~= 1.1",
    "sphinx>=3.2.1,<4",
    "recommonmark>=0.6.0,<0.7",
    "pytest-cov>=2.11.1",
    "ruff>=0.9",
]

[tool.uv]

[tool.uv.build-backend]
module-name = ["confme"]
module-root = ""

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[tool.pyright]
include = ["confme", "tests"]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
